Enchancments to the library to reflect better features of newer Paper versions.

+++ src/main/java/com/skyevel/universalscoreboard/api/UniversalScoreboardAPI.java
@@ -1,12 +1,14 @@
 package com.skyevel.universalscoreboard.api;
 
-import com.skyevel.universalscoreboard.feature.fastboard.FastBoard;
+import com.skyevel.universalscoreboard.ComponentUtil;
 import com.skyevel.universalscoreboard.workers.BoardWorker;
-import org.bukkit.Bukkit;
+import fr.mrmicky.fastboard.adventure.FastBoard;
+import net.kyori.adventure.text.Component;
 import org.bukkit.entity.Player;
 
-public class UniversalScoreboardAPI {
+import java.util.List;
 
+public class UniversalScoreboardAPI {
     /**
      * Get the FastBoard instance attached to a player.
      *
@@ -14,32 +16,106 @@ public class UniversalScoreboardAPI {
      * @return the scoreboard instance
      */
     public static FastBoard getBoard(Player player) {
-        return BoardWorker.getBoards().get(player.getName());
+        return BoardWorker.findFor(player);
     }
 
     /**
      * Set the title of a player's scoreboard.
      *
      * @param player the player
-     * @param title the title
+     * @param titleComponent the title
      */
-    public static void setBoardTitle(Player player, String title) {
-
-        getBoard(player).updateTitle(title);
-
+    public static void setBoardTitle(Player player, Component titleComponent) {
+        getBoard(player).updateTitle(titleComponent);
     }
+
+    // Adventure API - start
+    // Dan66ika77 changes start
 
     /**
      * Create the board instance for a player.
      *
      * @param player the player
      */
-    public static void createBoard(Player player) {
-
+    public static FastBoard createBoard(Player player) {
         FastBoard board = new FastBoard(player);
-        board.updateTitle(" ");
-        BoardWorker.getBoards().put(player.getName(), board);
+        board.updateTitle(Component.empty());
+        BoardWorker.getBoards().add(board);
+        return board;
     }
+
+    /**
+     * Create the board instance for a player with initial title.
+     *
+     * @param player the player
+     * @param initialTitleComponent the initial title
+     */
+    public static FastBoard createBoard(Player player, Component initialTitleComponent) {
+        FastBoard board = new FastBoard(player);
+        board.updateTitle(initialTitleComponent);
+        BoardWorker.getBoards().add(board);
+        return board;
+    }
+
+    /**
+     * Create the board instance for a player with initial lines.
+     *
+     * @param player the player
+     * @param initialTitleComponent the initial title
+     * @param initialLineComponents the initial lines
+     */
+    public static FastBoard createBoard(Player player, Component initialTitleComponent, List<Component> initialLineComponents) {
+        FastBoard board = new FastBoard(player);
+        board.updateTitle(initialTitleComponent);
+        board.updateLines(initialLineComponents);
+        BoardWorker.getBoards().add(board);
+        return board;
+    }
+
+    // Dan66ika77 changes end
+    // Adventure API - end
+
+    // Legacy API - start
+
+    /**
+     * Set the title of a player's scoreboard.
+     *
+     * @param player the player
+     * @param titleString the title
+     */
+    public static void setBoardTitle(Player player, String titleString) {
+        getBoard(player).updateTitle(ComponentUtil.formatToComponent(titleString));
+    }
+
+    /**
+     * Create the board instance for a player with initial title.
+     *
+     * @param player the player
+     * @param initialTitleComponent the initial title
+     */
+    public static FastBoard createBoard(Player player, String initialTitleComponent) {
+        FastBoard board = new FastBoard(player);
+        board.updateTitle(ComponentUtil.formatToComponent(initialTitleComponent));
+        BoardWorker.getBoards().add(board);
+        return board;
+    }
+
+    /**
+     * Create the board instance for a player with initial lines.
+     *
+     * @param player the player
+     * @param initialTitleComponent the initial title
+     * @param initialLineComponents the initial lines
+     */
+    public static FastBoard createBoard(Player player, String initialTitleComponent, List<String> initialLineComponents) {
+        FastBoard board = new FastBoard(player);
+        board.updateTitle(ComponentUtil.formatToComponent(initialTitleComponent));
+        board.updateLines(ComponentUtil.formatComponentList(initialLineComponents));
+        BoardWorker.getBoards().add(board);
+        return board;
+    }
+
+    // Legacy API - end
 
     /**
      * Safely delete the board instance of a player.
@@ -47,14 +123,7 @@ public class UniversalScoreboardAPI {
      * @param player the player
      */
     public static void destroyBoard(Player player) {
-
-        String playerName = player.getName();
-
-        FastBoard board = BoardWorker.getBoards().remove(playerName);
-
-        if(board != null) {
-            board.delete();
-        }
-
+        FastBoard board = BoardWorker.findFor(player);
+        if (board != null) board.delete();
     }
 }
